pyproject.toml
[project]
name = "clearcare-compliance"
version = "0.1.0"
description = "CMS Hospital Price Transparency validator (JSON + CSV Tall/Wide) with human-readable reports."
readme = "README_CLEARCARE.md"
requires-python = ">=3.10"
authors = [{name="ClearCare", email="dev@clearcare.example"}]
license = "MIT"
dependencies = [
  "jsonschema>=4.19",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "rich>=13.7",
  "python-dateutil>=2.9",
  "click>=8.1"
]

[project.scripts]
clearcare-validate = "clearcare_compliance.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["clearcare_compliance*"]

[tool.setuptools.package-data]
clearcare_compliance = ["schemas/*/*", "schemas/*/*/*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
